Split inline_parser.dart
